Moved TG actions to a separate unit
